requirements.txt
djangorestframework
django
drf-yasg